Small makefile change; updated udev rules
Makefile change includes a space so that the colons line up properly by default.

The `udev` rule is more substantial, as it transitions away from the `plugdev` method of allowing the user to change the backlight to what is defined by default in the Archlinux wiki (what the system is mostly tested on). This should fix an error where the utility is not usable without root permissions.